What I did during the lab
