+++ src/main/scala/com/mblund/todo/views/MyHaxEvent.scala
@@ -1,0 +1,14 @@
+package com.mblund.todo.views
+
+import scala.scalajs.js
+
+
+
+trait MyHaxEvent extends js.Object {
+  def preventDefault(): Unit = js.native
+}
+
+
+
+
+
