+++ src/main/scala/com/mblund/todo/views/View.scala
@@ -1,0 +1,13 @@
+package com.mblund.todo.views
+
+import org.scalajs.dom._
+
+trait View {
+  def el: Element
+
+  def removeItself() = {
+    if (el.parentNode != null) {
+      el.parentNode.removeChild(el)
+    }
+  }
+}
